requirements.txt
numpy
pandas
matplotlib.pyplot
matplotlib.image
cv2
random
os
sklearn.preprocessing
keras.preprocessing
tensorflow.keras.optimizers
PIL
keras.models
tensorflow.keras.layers
keras.layers
sklearn.model_selection